remove contrexx from framework
installation or rather downloading and copying the files has to be done
manually to keep all relevant files and ensure that the latest version
is used. as soon as comvation releases a permanent link for the latest
version a fix will be made to automatically download the cms through
a grunt task.