move header image from index to content templates
since the full width header is only used on the home page it has been
removed from the index template and put into the home template.

Some content pages do have header images but only over the width
of the main content area (not over the sidebar). These are put
in through a block category.