style the team page
Using nested grids in Susy proved to be a little more difficult
than it seemed. So for the teams their own content template
had to be created so that a block category can be used
to add team members.
Also basic layouting of the staff component has been put
in its own layout partial instead of something like
regions/_main.scss.