assign icon class to metanav search
instead of defining styles on the .icon-search or .icon-search-lines class
which can change with the different color configs an additional .icon class
has been added to the icon i which now can be styled independently of the
icon that is actually used. Also the metanav will probably end up as a
block container.