improve news teasers
finish up the styling for the news teaser component. Also needed
a small update to the styleguide example.
The news teasers are now styled nicely with SVG icons instead of
> for the links. Also the script has been written to make the
whole teaser clickable.
The news teasers have also been added to the template and are
now available on the page.