Add styles config for editor
Since the editor styles cannot be put into the customized folder we have to copy
them directly to the lib folder. It's very probable that we'll have to do the same
thing for the templates file. If so make sure to edit the task here created to just copy
the whole folder.