Template maintenance and new one
A new nested template has been created. Within this everything in it must be
contained in susy containers. This happened in preparation of two row sections.

Also a typo has been corrected in the index template and a block container category
has been added next to the sidebar. This makes it possible to add custom sidebars
or subnavs to pages without subpages.