Use hologram with bundler
The grunt-hologram module has been removed in favor of a direct shell
task integration. This makes it possible to run the command with `bundle
exec` instead of the general `hologram` command that is run from the
module previously used. As we need to specify the location of the configuration
file this has been added to `package.json`.