Add the rsync task
Will enable rsync deployment over SSH. Seems to work properly however cannot
be fully tested at this point in time since we need the password first.

Also is there a possibility to set up a key exchange so no password is needed
after initial connect? Or is there a way to pass the password from the
`secret.json` file?