fix(styles): add icons to editor styles
We need to manually import the icons svg for the editor styles so they show
up in the backend editor when they are applied. The easiest and dirtiest way
is to just css import the file but since it is the backend we need not worry
about performance that much.