feat(cms): remove thumb calls for news module
The news module should not be calling thumbnail images since we mostly only
use news images in a thumbnail form and therefore size them optimally.
This is done by editing the parseImageThumbnail function to ignore the
thumbnails. Also for the teasers we need an extra edit since there the
function is not called for an identical operation.