feat(server): add server config template
A basic .htaccess file is placed in the config folder to be used after successfully
developing a theme. After installing Contrexx the .htaccess is generated and can in
theory be overwritten then. However possible caching issues might lead you to want to
implement this at a later time when going live or testing in a live environment.