feat(build): optimize grunt config
We move the folder definitions outside of the package.json file and
into its own Gruntconfig.yml file. This helps us in that we have an
easier to read and use syntax in the yaml file and also cleans up
the package.json and moves data that does not necessarily belong
there out of it.