feat(cms): add .htaccess template
We add the template and the tasks to automatically copy it
to the root folder on every grunt run. We might want to
consider temporarily disabling this or setting it through
a flag to make sure we do not have to deal with too many
caching issues. On the other hand it is great to work in the
same exact environment the end user will experience.